more consistent New menu

+++ main/main_Praat.cpp
@@ -47,7 +47,7 @@ static void logo (Graphics graphics) {
 int main (int argc, char *argv []) {
 	try {
 		praat_setLogo (130.0, 80.0, logo);
-		praat_init (U"Praat", (unsigned int) argc, argv);
+		praat_init (U"Praat", argc, argv);
 		INCLUDE_LIBRARY (praat_uvafon_init)
 		INCLUDE_LIBRARY (praat_contrib_Ola_KNN_init)
 		praat_run ();
